Small front end change and CQRS

+++ AP.CityDataDemo.Domain/Entities/City.cs
@@ -6,11 +6,11 @@ public class City : IEntity
 {
     public int Id { get; set; }
     public string Name { get; private set; }
-    public int Population { get; private set; }
+    public long Population { get; private set; }
     public int CountryId { get; private set; }
     public Country Country { get; private set; } = null!;
 
-    public City(string name, int population, int countryId)
+    public City(string name, long population, int countryId)
     {
         if (string.IsNullOrWhiteSpace(name))
             throw new ArgumentException("City name cannot be null or empty.", nameof(name));
@@ -24,7 +24,7 @@ public class City : IEntity
         CountryId = countryId;
     }
 
-    public void UpdatePopulation(int newPopulation)
+    public void UpdatePopulation(long newPopulation)
     {
         if (newPopulation < 0)
             throw new ArgumentException("Population cannot be negative.", nameof(newPopulation));
